feat: implement lyrics module with SilkyWave animation and utility functions
